+++ docs/components/DocumentationPage.tsx
@@ -4,8 +4,6 @@ import some from 'lodash/some';
 import Router, { NextRouter } from 'next/router';
 import * as React from 'react';
 
-import { AppJSBanner } from './AppJSBanner.tsx';
-
 import * as Utilities from '~/common/utilities';
 import * as WindowUtils from '~/common/window';
 import DocumentationFooter from '~/components/DocumentationFooter';
@@ -62,10 +60,11 @@ class DocumentationPageWithApiVersion extends React.Component<Props, State> {
   componentDidMount() {
     Router.events.on('routeChangeStart', url => {
       if (this.layoutRef.current) {
-        window.__sidebarScroll =
-          this.getActiveTopLevelSection() !== this.getActiveTopLevelSection(url)
-            ? 0
-            : this.layoutRef.current.getSidebarScrollTop();
+        if (this.getActiveTopLevelSection() !== this.getActiveTopLevelSection(url)) {
+          window.__sidebarScroll = 0;
+        } else {
+          window.__sidebarScroll = this.layoutRef.current.getSidebarScrollTop();
+        }
       }
     });
     window.addEventListener('resize', this.handleResize);
@@ -144,14 +143,14 @@ class DocumentationPageWithApiVersion extends React.Component<Props, State> {
   private getActiveTopLevelSection = (path?: string) => {
     if (this.isReferencePath(path)) {
       return 'reference';
+    } else if (this.isEasPath(path)) {
+      return 'eas';
     } else if (this.isGeneralPath()) {
       return 'general';
     } else if (this.isFeaturePreviewPath(path)) {
       return 'featurePreview';
     } else if (this.isPreviewPath()) {
       return 'preview';
-    } else if (this.isEasPath(path)) {
-      return 'eas';
     } else if (this.isArchivePath()) {
       return 'archive';
     }
@@ -160,9 +159,9 @@ class DocumentationPageWithApiVersion extends React.Component<Props, State> {
   };
 
   render() {
-    const sidebarScrollPosition = process.browser ? window.__sidebarScroll : 0;
     const routes = this.getRoutes();
     const sidebarActiveGroup = this.getActiveTopLevelSection();
+    const sidebarScrollPosition = process.browser ? window.__sidebarScroll : 0;
 
     const sidebarElement = <Sidebar routes={routes} />;
     const headerElement = (
@@ -192,7 +191,6 @@ class DocumentationPageWithApiVersion extends React.Component<Props, State> {
     const pageContent = (
       <>
         {this.props.title && <H1>{this.props.title}</H1>}
-        <AppJSBanner />
         {this.props.children}
         {this.props.title && (
           <DocumentationFooter