+++ packages/expo-modules-core/android/src/main/cpp/types/FrontendConverter.h
@@ -12,8 +12,7 @@ namespace jsi = facebook::jsi;
 
 namespace expo {
 class JSIInteropModuleRegistry;
-
-class ExpectedType;
+class SingleType;
 
 /**
  * A base interface for all frontend converter classes - converters that cast jsi values into JNI objects.
@@ -237,7 +236,7 @@ private:
 class PrimitiveArrayFrontendConverter : public FrontendConverter {
 public:
   PrimitiveArrayFrontendConverter(
-    jni::local_ref<jni::JavaClass<ExpectedType>::javaobject> expectedType
+    jni::local_ref<jni::JavaClass<SingleType>::javaobject> expectedType
   );
 
   jobject convert(
@@ -262,5 +261,53 @@ private:
    * Converter used to convert array elements.
    */
   std::shared_ptr<FrontendConverter> parameterConverter;
+};
+
+/**
+ * Converter from js array object to [java.utils.ArrayList].
+ */
+class ListFrontendConverter : public FrontendConverter {
+public:
+  ListFrontendConverter(
+    jni::local_ref<jni::JavaClass<SingleType>::javaobject> expectedType
+  );
+
+  jobject convert(
+    jsi::Runtime &rt,
+    JNIEnv *env,
+    JSIInteropModuleRegistry *moduleRegistry,
+    const jsi::Value &value
+  ) const override;
+
+  bool canConvert(jsi::Runtime &rt, const jsi::Value &value) const override;
+private:
+  /**
+   * Converter used to convert array elements.
+   */
+  std::shared_ptr<FrontendConverter> parameterConverter;
+};
+
+/**
+ * Converter from js object to [java.utils.LinkedHashMap].
+ */
+class MapFrontendConverter : public FrontendConverter {
+public:
+  MapFrontendConverter(
+    jni::local_ref<jni::JavaClass<SingleType>::javaobject> expectedType
+  );
+
+  jobject convert(
+    jsi::Runtime &rt,
+    JNIEnv *env,
+    JSIInteropModuleRegistry *moduleRegistry,
+    const jsi::Value &value
+  ) const override;
+
+  bool canConvert(jsi::Runtime &rt, const jsi::Value &value) const override;
+private:
+  /**
+   * Converter used to convert values.
+   */
+  std::shared_ptr<FrontendConverter> valueConverter;
 };
 } // namespace expo